// lib/proto/client_token.ts
export interface ClientTokenRequest {
    /**
     * @generated from protobuf field: spotify.clienttoken.http.v0.ClientTokenRequestType request_type = 1;
     */
    requestType: ClientTokenRequestType;
    /**
     * @generated from protobuf oneof: request
     */
    request: {
        oneofKind: "clientData";
        /**
         * @generated from protobuf field: spotify.clienttoken.http.v0.ClientDataRequest client_data = 2;
         */
        clientData: ClientDataRequest;
    } | {
        oneofKind: "challengeAnswers";
        /**
         * @generated from protobuf field: spotify.clienttoken.http.v0.ChallengeAnswersRequest challenge_answers = 3;
         */
        challengeAnswers: ChallengeAnswersRequest;
    } | {
        oneofKind: undefined;
    };
}
/**
 * @generated from protobuf message spotify.clienttoken.http.v0.ClientDataRequest
 */
export interface ClientDataRequest {
    /**
     * @generated from protobuf field: string client_version = 1;
     */
    clientVersion: string;
    /**
     * @generated from protobuf field: string client_id = 2;
     */
    clientId: string;
    /**
     * @generated from protobuf oneof: data
     */
    data: {
        oneofKind: "connectivitySdkData";
        /**
         * @generated from protobuf field: spotify.clienttoken.data.v0.ConnectivitySdkData connectivity_sdk_data = 3;
         */
        connectivitySdkData: ConnectivitySdkData;
    } | {
        oneofKind: undefined;
    };
}
/**
 * @generated from protobuf message spotify.clienttoken.http.v0.ChallengeAnswersRequest
 */
export interface ChallengeAnswersRequest {
    /**
     * @generated from protobuf field: string state = 1;
     */
    state: string;
    /**
     * @generated from protobuf field: repeated spotify.clienttoken.http.v0.ChallengeAnswer answers = 2;
     */
    answers: ChallengeAnswer[];
}
/**
 * @generated from protobuf message spotify.clienttoken.http.v0.ClientTokenResponse
 */
export interface ClientTokenResponse {
    /**
     * @generated from protobuf field: spotify.clienttoken.http.v0.ClientTokenResponseType response_type = 1;
     */
    responseType: ClientTokenResponseType;
    /**
     * @generated from protobuf oneof: response
     */
    response: {
        oneofKind: "grantedToken";
        /**
         * @generated from protobuf field: spotify.clienttoken.http.v0.GrantedTokenResponse granted_token = 2;
         */
        grantedToken: GrantedTokenResponse;
    } | {
        oneofKind: "challenges";
        /**
         * @generated from protobuf field: spotify.clienttoken.http.v0.ChallengesResponse challenges = 3;
         */
        challenges: ChallengesResponse;
    } | {
        oneofKind: undefined;
    };
}
/**
 * @generated from protobuf message spotify.clienttoken.http.v0.TokenDomain
 */
export interface TokenDomain {
    /**
     * @generated from protobuf field: string domain = 1;
     */
    domain: string;
}
/**
 * @generated from protobuf message spotify.clienttoken.http.v0.GrantedTokenResponse
 */
export interface GrantedTokenResponse {
    /**
     * @generated from protobuf field: string token = 1;
     */
    token: string;
    /**
     * @generated from protobuf field: int32 expires_after_seconds = 2;
     */
    expiresAfterSeconds: number;
    /**
     * @generated from protobuf field: int32 refresh_after_seconds = 3;
     */
    refreshAfterSeconds: number;
    /**
     * @generated from protobuf field: repeated spotify.clienttoken.http.v0.TokenDomain domains = 4;
     */
    domains: TokenDomain[];
}
/**
 * @generated from protobuf message spotify.clienttoken.http.v0.ChallengesResponse
 */
export interface ChallengesResponse {
    /**
     * @generated from protobuf field: string state = 1;
     */
    state: string;
    /**
     * @generated from protobuf field: repeated spotify.clienttoken.http.v0.Challenge challenges = 2;
     */
    challenges: Challenge[];
}
// ...
export interface Challenge {
    /**
     * @generated from protobuf field: spotify.clienttoken.http.v0.ChallengeType type = 1;
     */
    type: ChallengeType;
    /**
     * @generated from protobuf oneof: parameters
     */
    parameters: {
        oneofKind: "clientSecretParameters";
        /**
         * @generated from protobuf field: spotify.clienttoken.http.v0.ClientSecretParameters client_secret_parameters = 2;
         */
        clientSecretParameters: ClientSecretParameters;
    } | {
        oneofKind: "evaluateJsParameters";
        /**
         * @generated from protobuf field: spotify.clienttoken.http.v0.EvaluateJSParameters evaluate_js_parameters = 3;
         */
        evaluateJsParameters: EvaluateJSParameters;
    } | {
        oneofKind: "evaluateHashcashParameters";
        /**
         * @generated from protobuf field: spotify.clienttoken.http.v0.HashCashParameters evaluate_hashcash_parameters = 4;
         */
        evaluateHashcashParameters: HashCashParameters;
    } | {
        oneofKind: undefined;
    };
}
// ...
export interface ChallengeAnswer {
    /**
     * @generated from protobuf field: spotify.clienttoken.http.v0.ChallengeType ChallengeType = 1 [json_name = "ChallengeType"];
     */
    challengeType: ChallengeType;
    /**
     * @generated from protobuf oneof: answer
     */
    answer: {
        oneofKind: "clientSecret";
        /**
         * @generated from protobuf field: spotify.clienttoken.http.v0.ClientSecretHMACAnswer client_secret = 2;
         */
        clientSecret: ClientSecretHMACAnswer;
    } | {
        oneofKind: "evaluateJs";
        /**
         * @generated from protobuf field: spotify.clienttoken.http.v0.EvaluateJSAnswer evaluate_js = 3;
         */
        evaluateJs: EvaluateJSAnswer;
    } | {
        oneofKind: "hashCash";
        /**
         * @generated from protobuf field: spotify.clienttoken.http.v0.HashCashAnswer hash_cash = 4;
         */
        hashCash: HashCashAnswer;
    } | {
        oneofKind: undefined;
    };
}
/**
 * @generated from protobuf message spotify.clienttoken.http.v0.ClientTokenBadRequest
 */
export interface ClientTokenBadRequest {
    /**
     * @generated from protobuf field: string message = 1;
     */
    message: string;
}
/**
 * @generated from protobuf enum spotify.clienttoken.http.v0.ClientTokenRequestType
 */
export enum ClientTokenRequestType {
    /**
     * @generated from protobuf enum value: REQUEST_UNKNOWN = 0;
     */
    REQUEST_UNKNOWN = 0,
    /**
     * @generated from protobuf enum value: REQUEST_CLIENT_DATA_REQUEST = 1;
     */
    REQUEST_CLIENT_DATA_REQUEST = 1,
    /**
     * @generated from protobuf enum value: REQUEST_CHALLENGE_ANSWERS_REQUEST = 2;
     */
    REQUEST_CHALLENGE_ANSWERS_REQUEST = 2
}
/**
 * @generated from protobuf enum spotify.clienttoken.http.v0.ClientTokenResponseType
 */
export enum ClientTokenResponseType {
    /**
     * @generated from protobuf enum value: RESPONSE_UNKNOWN = 0;
     */
    RESPONSE_UNKNOWN = 0,
    /**
     * @generated from protobuf enum value: RESPONSE_GRANTED_TOKEN_RESPONSE = 1;
     */
    RESPONSE_GRANTED_TOKEN_RESPONSE = 1,
    /**
     * @generated from protobuf enum value: RESPONSE_CHALLENGES_RESPONSE = 2;
     */
    RESPONSE_CHALLENGES_RESPONSE = 2
}
// ...
class ClientTokenRequest$Type extends MessageType<ClientTokenRequest> {
    constructor() {
        super("spotify.clienttoken.http.v0.ClientTokenRequest", [
            { no: 1, name: "request_type", kind: "enum", T: () => ["spotify.clienttoken.http.v0.ClientTokenRequestType", ClientTokenRequestType] },
            { no: 2, name: "client_data", kind: "message", oneof: "request", T: () => ClientDataRequest },
            { no: 3, name: "challenge_answers", kind: "message", oneof: "request", T: () => ChallengeAnswersRequest }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message spotify.clienttoken.http.v0.ClientTokenRequest
 */
export const ClientTokenRequest = new ClientTokenRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ClientDataRequest$Type extends MessageType<ClientDataRequest> {
    constructor() {
        super("spotify.clienttoken.http.v0.ClientDataRequest", [
            { no: 1, name: "client_version", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "client_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "connectivity_sdk_data", kind: "message", oneof: "data", T: () => ConnectivitySdkData }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message spotify.clienttoken.http.v0.ClientDataRequest
 */
export const ClientDataRequest = new ClientDataRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ChallengeAnswersRequest$Type extends MessageType<ChallengeAnswersRequest> {
    constructor() {
        super("spotify.clienttoken.http.v0.ChallengeAnswersRequest", [
            { no: 1, name: "state", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "answers", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => ChallengeAnswer }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message spotify.clienttoken.http.v0.ChallengeAnswersRequest
 */
export const ChallengeAnswersRequest = new ChallengeAnswersRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ClientTokenResponse$Type extends MessageType<ClientTokenResponse> {
    constructor() {
        super("spotify.clienttoken.http.v0.ClientTokenResponse", [
            { no: 1, name: "response_type", kind: "enum", T: () => ["spotify.clienttoken.http.v0.ClientTokenResponseType", ClientTokenResponseType] },
            { no: 2, name: "granted_token", kind: "message", oneof: "response", T: () => GrantedTokenResponse },
            { no: 3, name: "challenges", kind: "message", oneof: "response", T: () => ChallengesResponse }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message spotify.clienttoken.http.v0.ClientTokenResponse
 */
export const ClientTokenResponse = new ClientTokenResponse$Type();
// @generated message type with reflection information, may provide speed optimized methods
class TokenDomain$Type extends MessageType<TokenDomain> {
    constructor() {
        super("spotify.clienttoken.http.v0.TokenDomain", [
            { no: 1, name: "domain", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message spotify.clienttoken.http.v0.TokenDomain
 */
export const TokenDomain = new TokenDomain$Type();
// @generated message type with reflection information, may provide speed optimized methods
class GrantedTokenResponse$Type extends MessageType<GrantedTokenResponse> {
    constructor() {
        super("spotify.clienttoken.http.v0.GrantedTokenResponse", [
            { no: 1, name: "token", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "expires_after_seconds", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 3, name: "refresh_after_seconds", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 4, name: "domains", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => TokenDomain }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message spotify.clienttoken.http.v0.GrantedTokenResponse
 */
export const GrantedTokenResponse = new GrantedTokenResponse$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ChallengesResponse$Type extends MessageType<ChallengesResponse> {
    constructor() {
        super("spotify.clienttoken.http.v0.ChallengesResponse", [
            { no: 1, name: "state", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "challenges", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => Challenge }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message spotify.clienttoken.http.v0.ChallengesResponse
 */
export const ChallengesResponse = new ChallengesResponse$Type();
// ...
export const Challenge = new Challenge$Type();
// ...
export const ChallengeAnswer = new ChallengeAnswer$Type();